refactor: change header component
