+++ src/components/header/index.ts
@@ -1,1 +1,7 @@
 export * from './header';
+
+export { LogoutButton } from './logout-button/logout-button';
+export { ManagedList } from './managed-list/managed-list';
+export { ManagedNavigationButtons } from './managed-navigation-buttons/managed-navigation-buttons';
+export { RegisterLoginButtons } from './register-login-buttons/register-login-buttons';
+export { Sections } from './sections/sections';
