+++ src/components/header/navigation.constants.ts
@@ -1,0 +1,4 @@
+const sectionsLabels = ['about', 'catalog', 'cart'];
+const buttonsLabels = ['registration', 'login'];
+
+export { buttonsLabels, sectionsLabels };
